feat: Category enum ALL 도메인 추가 (#171)
