+++ src/main/java/com/smunity/server/global/common/entity/enums/Category.java
@@ -10,6 +10,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public enum Category {
 
+    ALL("전체"),
     MAJOR_ADVANCED("1전심"),
     MAJOR_OPTIONAL("1전선"),
     CULTURE("교양"),
